Create README.md
I have developed a platform where items that need to be sold urgently will be sold. For example, a person will go on vacation and needs money urgently and there is a product that he wants to get rid of. Consider a platform that allows one-day sales. If you want the money for the product sold to be deposited within one day, an 8% commission will be deducted from the product sold. If you want the money to be deposited immediately, a commission of 12% of the sales price of the product will be deducted. In addition, the price of the product will increase by 0.3% every hour during a day's filling period. I create animated graphical comments, diagram representation and HTML. I designed a system that is supported and will appear in your web browser with javascript/typescript support.